+++ smtpdfix/__init__.py
@@ -1,7 +1,7 @@
 from .authenticator import Authenticator
 from .certs import generate_certs
 from .controller import AuthController
-from .fixture import smtpd
+from .fixture import SMTPDFix, smtpd
 from .handlers import AuthMessage
 from .smtp import AuthSMTP
 
@@ -12,4 +12,5 @@ __all__ = (
     "AuthSMTP",
     "generate_certs",
     "smtpd",
+    "SMTPDFix",
 )